Add files via upload
Hi Kumarvel Sir , Seesha Sir and Neela Sir

Check the attachment for the images.

Image 1 : Tells the corresponding content throws 500 internal server Error.

Image 2: Check the “Status” in the response of the content its been updated by 
roleId . Which will be done only throught ML Alogirthm because loopback doesnt
update any status.

Image 3 :  For the roleId that is update through the 
ML Algorithm is not available in the RDS.

So 500 error is thrown . Check the ML Algorithm its updating the 
RoleId Wrongly .

This is what Naren conveyed through a mail on the other day that there is some issue 
with the ML ALgorithm validation.  

Fortunately finally the issue is found make the correction the ML Side and check the issue again .

Since the ML code is freezed long back it could be good to refractor and retest the ML code once again.